requirements.txt
anthropic
python-dotenv
requests
feedparser
mechanicalsoup
rich
python-docx